pyproject.toml
[tool.poetry]
name = "observability-mtl-instrument"
version = "0.1.0"
description = ""
authors = ["SergioNascimento07 <sergioricjr7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2023.3.post1"
opentelemetry-instrumentation-logging = "^0.42b0"
prometheus-client = "^0.18.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-otlp-proto-http = "^1.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.12.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.4.8"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks] 
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest --cov -s -x -vv"
post_test = "coverage html"
